tweak mssql data type testing structure and decode tinyint as short
