+++ vertx-mssql-client/src/test/java/io/vertx/mssqlclient/data/MSSQLNotNullableDataTypeTestBase.java
@@ -1,0 +1,4 @@
+package io.vertx.mssqlclient.data;
+
+public abstract class MSSQLNotNullableDataTypeTestBase extends MSSQLFullDataTypeTestBase {
+}
